refactor(core): move auth and others to core
