+++ src/app/@core/auth/auth.service.ts
@@ -1,24 +1,16 @@
-import { Injectable } from '@angular/core';
+import { Injectable, inject } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { User, CreateUserDto, Credentials } from './user';
+import { User, CreateUserDto, Credentials, UserService } from '../user';
 import { environment } from 'src/environments/environment';
-import { BehaviorSubject, tap } from 'rxjs';
+import { tap } from 'rxjs';
 
 @Injectable()
 export class AuthService {
-  private _user = new BehaviorSubject<Partial<User> | null>(null);
-  readonly user$ = this._user.asObservable();
-
   private readonly ACCESS_TOKEN = 'accessToken';
-  private readonly USER = 'user';
-  constructor(private httpClient: HttpClient) {
-    const userLocalStorage = localStorage.getItem(this.USER);
-    if (userLocalStorage) {
-      this.user = JSON.parse(userLocalStorage);
-    }
-  }
 
-  // TODO: Store in Ionic Storage (IndexedDB)
+  private readonly http = inject(HttpClient);
+  private readonly userService = inject(UserService);
+
   get accessToken() {
     return localStorage.getItem(this.ACCESS_TOKEN) ?? '';
   }
@@ -27,17 +19,8 @@ export class AuthService {
     localStorage.setItem(this.ACCESS_TOKEN, token);
   }
 
-  get user() {
-    return this._user.getValue();
-  }
-
-  set user(user: Partial<User> | null) {
-    this._user.next(user);
-    localStorage.setItem(this.USER, JSON.stringify(user));
-  }
-
   register({ email, fullName, passwords }: CreateUserDto) {
-    return this.httpClient
+    return this.http
       .post<User>(`${environment.api}/auth/register`, {
         email,
         fullName,
@@ -51,7 +34,7 @@ export class AuthService {
   }
 
   login({ email, password }: Credentials) {
-    return this.httpClient
+    return this.http
       .post<User>(`${environment.api}/auth/login`, {
         email,
         password,
@@ -65,6 +48,6 @@ export class AuthService {
 
   private setCredentials({ token, ...user }: User) {
     this.accessToken = token;
-    this.user = user;
+    this.userService.user = user;
   }
 }
