+++ src/app/@core/auth/index.ts
@@ -1,3 +1,2 @@
 export { AuthModule } from "./auth.module";
 export { AuthService } from "./auth.service";
-export { credentialsValidations, userValidations } from "./user.validations";
