+++ src/app/@core/services/storage/storage.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 
-@Injectable({ providedIn: 'root' })
+@Injectable()
 export class StorageService {
   constructor() {}
 }
