+++ src/app/@core/user/index.ts
@@ -1,0 +1,3 @@
+export * from '../user/user.validations';
+export * from './user.service';
+export * from './user';
