+++ src/app/app.module.ts
@@ -6,17 +6,15 @@ import { IonicModule, IonicRouteStrategy } from '@ionic/angular';
 
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
-import { AuthModule } from '@auth/auth.module';
-import { HttpClientModule } from '@angular/common/http';
+import { CoreModule } from './@core/core.module';
 
 @NgModule({
   declarations: [AppComponent],
   imports: [
     BrowserModule,
-    IonicModule.forRoot(),
     AppRoutingModule,
-    HttpClientModule,
-    AuthModule,
+    IonicModule.forRoot(),
+    CoreModule,
   ],
   providers: [{ provide: RouteReuseStrategy, useClass: IonicRouteStrategy }],
   bootstrap: [AppComponent],
