+++ tsconfig.json
@@ -19,18 +19,15 @@
     "target": "es2022",
     "module": "es2020",
     "paths": {
-      "@pages/*": [
-        "src/app/pages/*"
-      ],
       "@shared/*": [
         "src/app/@shared/*"
       ],
-      "@auth/*": [
-        "src/app/@auth/*"
+      "@core/*": [
+        "src/app/@core/*"
       ],
-      "@storage/*": [
-        "src/app/@storage/*"
-      ],
+      "@pages/*": [
+        "src/app/pages/*"
+      ]
     },
     "lib": [
       "es2018",
